+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Option to reset end device payload formatters in the Console.
+- Service discovery using DNS SRV records for external Application Server linking.
 
 ### Changed